2016-11-14 test connect

+++ Yingerbao/src/com/aizi/yingerbao/breath/BabyBreath.java
@@ -1,7 +1,7 @@
 package com.aizi.yingerbao.breath;
 
 public class BabyBreath {
-    public int mBreathTime;
+    public long mBreathTime; // 与上次呼吸的时间间隔，单位毫秒
     public int mBreathValue;
     public int mBreathFreq;
 }
